requirements.txt
PyYAML==5.3.1
xmltodict==0.12.0
requests==2.26.0
compile==1.0.3
SQLAlchemy==1.4.27
pyinstaller==4.10
paramiko==2.10.4
cryptography==36.0.2
APScheduler==3.9.1
pymysql==1.0.2
pg8000==1.29.4
pymssql==2.2.7
uvicorn==0.20.0
autopep8==2.0.0
fastapi==0.88.0